// src/Core/Enums/SpklActions.ts
export enum SpklActions {
    instrument = 'instrument',
    deployPlugins = 'plugins',
    deployWorkflows = 'workflow',
    getWebResources = 'get-webresources',
    deployWebResources = 'webresources',
    downloadWebResources = 'download-webresources',
    generateEntities = 'earlybound',
    exportSolution = 'unpack',
    importSolution = 'import'
}